Add a setting to change the web root
The code currently treats the workspace root as the web root. If the ColdFusion/Lucee server uses a subfolder instead, then we should allow setting that as the web root.

Without this, component names that are relative to the web root won't be found. A workaround is adding custom mappings for each child folder of the web root. These are mappings that are not needed on the server.